facebook_auto_reporter/requirements.txt
selenium==4.15.0
openai==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.4
python-dotenv==1.0.0
fake-useragent==1.4.0
Flask==2.3.3
gunicorn==21.2.0
flask-socketio==5.3.6